Add files via upload
Description:
This program finds and displays all prime numbers from 0 to a number entered by the user. The user is prompted to input a number, and the program checks each number from 2 up to the input number to determine if it is a prime number. If the number is prime, it will be printed